pyproject.toml
[project]
name = "backend-app"
version = "0.1.0"
description = ""
authors = [
    {name = "=",email = "kumarmridul91@gmail.com"}
]
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.129.0,<0.130.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "pydantic-settings (>=2.13.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "greenlet (>=3.3.1,<4.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "app"}]

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "aiosqlite (>=0.22.1,<0.23.0)"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"


